Add additional node indexes
Secure RBAC, along with numerous periodics, and even some common API
access patterns heavily access the ironic database and sometimes cause
queries across multiple columns to match nodes to return.

None of this is bad, but what is bad is we didn't have indexes on these
columns.

This change adds docs, and the schema upgrade to create the column
indexes, and a release note to provide more visible documentation
for operators.

It must be stressed that this does *not* improve query times
when all records are asked for on a database connection.

Also adds an upgrade check in to raise a warning for operator
visibility.

Story: 2008863
Task: 42392
Change-Id: I76821c032180a58d0f99d31110fbe0f844c0cb3f