Ironic index docs/command check backport
Secure RBAC, along with numerous periodics, and even some common API
access patterns heavily access the ironic database and sometimes cause
queries across multiple columns to match nodes to return.

This change is largely intended to provide the upgrade check
call as well as the documentation in the Wallaby release
of ironic so operators can locate it and add the indexes prior
to the upgrade to ironic 18.0.

Note: This change includes the content from
Ifa9da65dc8df5bf9a369d6faeab310386dfd944a which includes a fix for
the status command and the upgrade script as to properly identify and
raise visibility, where the original change to the status command
failed to include the actual value in the field to cause the
standard execution by the upgrade script. Also includes the same
upgrade check handling for ironic-status command in the stable branch
devstack plugin as this change does raise an informational warning
which is non-blocking to ironic's use on this branch to signal
that action may be advisable.

Story: 2008863
Task: 42392
Change-Id: I76821c032180a58d0f99d31110fbe0f844c0cb3f
(cherry picked from commit 205a8b3037b62cbae7f5385f4d9b9404a702414e)