Provide an option to not cache bootable iso ramdisks
The behavior when a bootable iso is provided behind an http server is
to download and serve the image from the conductor; the image is
removed only when the node is undeployed.
In certain cases, for example on large deployments, this could cause
undesired behaviors, like the conductor nodes running out of disk storage.
We provide an option to avoid this event directly using the original url of
the bootable iso that is then served directly by the external http server
where it is hosted.

Change-Id: I6ecbe8ae4c775f5a39b68ff698bde655795e5db7