Change UEFI ipxe bootloader default
The bootloader default for ipxe, previously set to ipxe.efi
was, while consistent with the interface naming, a likely
setting that had to be overridden by most deployments,
as very little hardware can make use of ipxe.efi for network
booting, where as snponly.efi is realistically what should
be used in nearly all cases with the exception of some
extremely early UEFI supporting hardware, which may
happen to work with ipxe.efi *if* support for the network
card happens to be compiled in.

Note: This was likely never observable in CI as CI leverages
firmware with integrated iPXE binaries.

Change-Id: I808cf92c98339aa447b4799900066c310e15fd5b