Suppress policy deprecation and default change warnings
As part of the policy defaults refresh work, all of the defaults for
policies were changed and warnings about it are being emitted
constantly in the logs. We can suppress these for now until we are
enforcing new defaults [1].

This also suppresses policy deprecation warnings during tests to avoid
filling the console with warning messages.

Related to blueprint policy-defaults-refresh

[1]https://docs.openstack.org/nova/latest/configuration/config.html#oslo_policy.enforce_new_defaults

Change-Id: I4a169b42d516a7ae2adcee9b69d71723fdefcd95