Add reno and reset legacy policy deprecation expectation
Added a reno to support
https://review.opendev.org/c/openstack/ironic/+/799120
and ultimately removed the "anticipated to be removed in the Xena
development cycle" warnings on legacy policy support as the effort
is taking longer than the cross-team effort anticipated.

Notes added to the code highlight that we need to update the
warnings at that time to ensure we appropriately set expectations
as to when the legacy policies will be removed.

Change-Id: Ie23788abaa61742cdea208fb0cf026bdecaab525