Avoid double file removal in create_boot_iso
The images.fetch call removes the provided destination file on error.
The create_boot_iso function calls it with a temporary file, which
results in another attempt to remove it. This attempt fails and masks
the actual exception (e.g. validation failure).

I doubt this problem is exposed to end users, but fixing it anyway.

Change-Id: I77ef1cabbcea9808771613fa37b27978fd0ce770