Fix subscription vendor passthru
* Instead of specifying the fields we want to oomit from create/get
subscription, now we are specifying the fields we want to be returned
(This will avoid problems when vendors are using different versions)

* When validating create_subscription are requiring that Context and
Protocol are present, use the default value from Redfish in case
they are not present in the Body request.

* When trying to delete a subscription that doesn't exist Ironic
returns 500 (default code from IronicException), this commit updates
the code to 404 to show that the subscription doesn't exist.

Change-Id: I81907be1ebc293118f5ffde4fd4d0485ade390df
(cherry picked from commit a8bbfae5aaf66284bc2a060d7a90ac9384eeb34d)