Allow initial versions to not be created yet
The database schema upgrade check had support for an
explicit list of known versions to handle the upgrade.

who knew!

Anyhow, we haven't used it in three years since the addition,
and it seems to make more sense to just be able to indicate
"we know initial field versions may not be able to be retrieved
and act accordingly". As such, when no table is found, the
pre-upgrade version check execution will continue onward fearlessly!

Call it a spiritual successor to Ibcf0728bc5d1b0cbdd78796526f9c93fc99e8c08

Change-Id: Icae5162c2501b0d1217ad0e6ee34ebef40e95204