Revert "Allow reboot to hard disk following iso ramdisk deploy."
This reverts commit 3bad548ce3d34207aa70fde8342082fe243bee31
based upon the comments on https://review.opendev.org/c/openstack/ironic/+/801343 which
was the stable/wallaby backport for this change
where Derek points out that the change was redundant and
Dmitry asks if the patch should be reverted.

This change is a partial revert where the documentation
changed in the original commit was actually correct all
along.

Change-Id: I4bc9dbeb334f176b7c8ed4e4b5ec36affc17e9cd