Fix in-band cleaning for ramdisk and anaconda deploy
We have implemented the cleaning prepare/tear_down, but haven't
implemented fetching/running in-band clean steps. This change moves
the cleaning logic from AgentDeployMixin to AgentBaseMixin, where it
arguably belongs.

In a follow-up patch I'm planning to reduce the number of mix-ins we
currently have, but that won't be backportable.

Change-Id: Ibc5610b14cea487d26191249e5c0333fdcd4b914