Always update cache for HTTP images if Last Modified is unknown
Currently we default to assuming the cache is up-to-date. This is likely
wrong. Normal web servers provide Last Modified for files they serve.
If it is absent, chances are high the image is served by some sort
of a dynamic service, which may modify the URL on fly.

In any case, always updating the image is a safer choice.

Change-Id: I0548db14a97638d26ebb687e8f47f1b295d1f774
(cherry picked from commit 94a1560d31b72052a94c0a4c07db2f9cf957ff26)