[iRMC] Avoid repeatedly resuming clean after creating raid configuration
Fixed the bug of repeated resume cleaning due to the value of `fgi_status` not being updated correctly when obtaining the
RAID configuration status of the node managed by the `irmc` hardware type.

```
Unexpected error when processing next clean step. TypeError: 'NoneType' object is not subscriptable
```

This `NoneType` error occurs because ironic resumes clean without waiting for IPA to get the clean steps, and then it tries to read the
clean steps which should be a list but actually is none.
During auto clean, resume clean should be triggered by IPA, but in this case, it is triggered by a iRMC periodic task which checks
the progress of running raid config.
This error does not occur every time, sometime raid can be configured, auto clean can complete without error and the node can be deployed successfully.

Story: #2009206
Task: #43265

Signed-off-by: Zhou Hao <zhouhao@fujitsu.com>
Change-Id: I5a1c5708bdc1709e928f0faf7e18396e260dc551