Dial back gate job memory allocation
Observed an OOM incident causing
ironic-tempest-ipa-partition-pxe_ipmitool to fail.

One vm started, the other seemed to try to start twice, but both times
stopped shortly into the run and the base OS had recorded in it an OOM
failure.

It appears the actual QEMU memory footprint being consumed when
configured at 3GB is upwards of 4GB, which obviously is too big to
fit in our 8GB VM instance.

Dialing back slightly, in hopes it stabilizes the job.

Change-Id: Id8cef722ed305e96d89b9960a8f60f751f900221