Add missing mode setting on pxe created directories
Two calls to fileutils.ensure_tree were missing the follow-up chmod
for CONF.pxe.dir_permission. This is fixed with a local ensure_tree
function which creates the directory with the appropriate mode.

The configuration help for the permissions attributes clarify that the
system default umask is masked out on the supplied value.

Change-Id: I124d2ba09b0fc78b284c1ae871ca5a16fc44b8c9