Refactor: move base_iso handling from create_boot_iso
This is basically a no-op and does not belong there. Most of base_iso
handling happens in _prepare_iso_image, which has been restructured to
avoid redundant actions when boot_iso is provided.

Remove TODOs to rebuild base_iso. There are no immediate plans to
implement it, and I'm not even sure how doable it is (e.g. for CoreOS).

Change-Id: I9a0b955ec22d39b6a31494bd2593fec486f6d8a9