Do not append filename parameter to image URL when using local file
Currently Ironic always adds a filename parameter to virtual media
image URL, for example:
http://localhost/redfish/boot.iso?filename=file.iso. This is useful
on certain BMCs which require a ".iso" suffix in the image URL
especially when using swift as the image source.
This should not be neccessary with file based virtual media images
and adding this parameter can cause issues in BMCs which do not
accept special characters such as "=" and "?" in virtual image URL.
This change alters Ironic's behaviour to only add the filename
parameter if swift backing store is used for the image.

Story: 2009278
Task: 43547

Change-Id: I3338bdb6a24d9695ce32d205221183b9b876bcc8