[iRMC] Convert the type of irmc_port to int
When using the node managed by the `irmc` hardware type,
if the port number of the bmc address is manually specified,
the following error will occur:

```
Value '443' is not supported for 'irmc_port'
```

However iRMC supports 80 and 443 ports, and default to 443.
The reason for this error is that the type of irmc_port is not converted to int.

Story: #2009671
Task: #43915

Signed-off-by: Zhou Hao <zhouhao@fujitsu.com>
Change-Id: I4e9274de09758fdb468382a7f88298a279f43e92