CI: Get tinyipa build working for CI usage
The qemu git server has updated it's ssl certificate, and can no longer
be verified by the software libraries and embedded tinycore data for
the 10.x release. Unfortunately, we have CI hard coded to an older
release and we cannot release a newer, older version of the tinycore
10.x supporting packages.

So, for our CI purposes, we just need a working build, so that
is all this patch attempts to do, get our CI in a working state
for the older branches.

It is an awful change. None of us like it, but we're stuck and
this is the only path forward short of completely abandoning
the use of the software builds on these branches.

Change-Id: Ia4434e9d2e4df49b26c33fcf371d821e0d44d6b7