Move place_loaders_for_boot to boot driver __init__
Host preparation file writing is already done by the __init__ method
of iPXEBoot. This change moves place_loaders_for_boot calls to
iPXEBoot and PXEBoot to be consistent, and to only write the files
when that driver is enabled.

This will mean multiple writes of the same file when subclasses of
these drivers are also enabled, but this overhead will be negligible.

Change-Id: I7e17f4d1a54cd6c5d1a4bf006a0d42db8d123a46