Explicitly pin upper sushy, reduce job count
Sushy 4.0.0 has breaking changes in it which prevent usage
with this version of ironic. While the global upper requirements
is unlikely ever to attempt to increment to 4.0.0, the explicit
requirements invocation, like what is in place for running
driver unit tests, was trying to pull in 4.0.0 resulting in test
failures.

Also explicitly fix the upper constraints for tox.ini to the
xena release branch, which was from the development cycle which
resulted in this bugfix branch.

And also reduces the overall set of jobs for the bugfix branch down
to more a needful set of jobs as this branch is intended for bugfixes
to be tracked on the 18.1 release.

Change-Id: Ief828f139bac9ba00f3250d8ddf4fd3d84a6fa7a
(cherry picked from commit b3a835bd2e1f20215d6e82ee302daa6db3dd2453)