Explicitly pin upper sushy version
Sushy 4.0.0 has breaking changes in it which prevent usage
with this version of ironic. While the global upper requirements
is unlikely ever to attempt to increment to 4.0.0, the explicit
requirements invocation, like what is in place for running
driver unit tests, was trying to pull in 4.0.0 resulting in test
failures.

Change-Id: Ief828f139bac9ba00f3250d8ddf4fd3d84a6fa7a