Inspector: better error message on DiscoveryFailure
The keystoneauth's error message on connection problems is very
confusing:

    Version requested but version discovery document was not found and
    allow_version_hack was False

While I'm trying to improve the situation in [1], it still won't be
perfect. This patch adds some context to DiscoveryFailure.

We should probably do the same for other services, but inspector is
particularly affected since it can be used standalone.

[1] https://review.opendev.org/c/openstack/keystoneauth/+/825540

Change-Id: Ibd41580d1389c7584639bf000e330fec5ee1b73d