Fix Train CI
CI for stable/train presently fails on:

./ironic/drivers/modules/drac/power.py:174:22: F821 undefined name '_'

Which was used for translations, but translations are not used in the
file, as such stripped away the _ method invocation.

Additionally, we have some semi-random job failures due to dual jobs
running. In particular we fixed this in the Ussuri branch previously
in Iea06632feee59714238f3b5b144bdadb733a0c35, but this change also
applies this, and comments out jobs which we know are no longer
functional or able to be run due to aspects outside of our control.

Change-Id: I98d0c92d17e78aa5624f969710090f668df8eb0f