Suppress Bandit B509 in snmp.py
Bandit's B509 (SNMPv3 no authentication/privacy) check is
triggered by our line wrapping, as the check looks at the
number of arguments, and triggers if it only parses 3 arguments.

In Ironic's case, we pass five arguments. Presumably, this is due
to the line wrapping. Ultimately this bandit check is very new
having only been added with in the last few months.

Change-Id: Idaee72a53a1f33bcab940c6ebe1e7c113a1cb1ce