Deprecate syslinux
Syslinux is a functionally abandoned Legacy BIOS boot mode bootloader
which has not seen updates since 2019, and is starting to see
discussion amongst linux distributions to remove explicit support
and packaging for Syslinux. Syslinux's relevance is also disappearing
as UEFI booting is becoming the standard. While syslinux did go ahead
and ensure their bootloader *could* be built and support UEFI,
distributions also didn't uniformly adopt packaging and support for
this bootloader.

This change proposes to deprecate it and notates the areas in which
functionality is deprecated.

Change-Id: Ic52007fa4f207561d282eb5ae54273885c0ab0c0